Backend/requirements.txt
click==8.1.3
Flask==2.2.2
Flask-SQLAlchemy==3.0.2
greenlet==3.0.3
gunicorn==20.1.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
psycopg2-binary==2.9.9
SQLAlchemy==1.4.46
Werkzeug==2.2.2
PyJWT==2.3.0
requests==2.26.0
Flask-JWT-Extended==4.5.2
Flask-Bcrypt==1.0.1
flask-cors==4.0.0


